backend_generative-AI-for-chat-transcription/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.46.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decouple==0.0.7
distro==1.9.0
docker==7.1.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
HyperPyYAML==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
lightning_sdk==0.2.22
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.83.0
openai-whisper==20240930
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
primePy==1.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.2.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==78.1.1
setuptools-rust==1.11.1
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.2
tqdm==4.67.1
transformers==4.52.4
triton==3.3.0
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
yarl==1.20.0
